fixed tsc build command, updated follow up titles

+++ src/config/discord/index.ts
@@ -1,10 +1,9 @@
-import { GatewayIntentBits, Client, Partials} from 'discord.js';
+import { GatewayIntentBits, Client, Partials, Collection, InteractionType, Events } from 'discord.js';
 import { SlashCommandStringOption, SlashCommandBuilder } from '@discordjs/builders';
-import discordModals, {showModal } from 'discord-modals';
 import { cmdArr } from './commands';
 import getError from '../../utils/get-error';
 import embedBuilder from '../../utils/discord/embed';
-import modalBuilder from '../../utils/discord/modal';
+import loginModalBuilder from '../../utils/discord/modal';
 import { getFollowCmd, getSingleCmd, helpCmd } from '../../interactions';
 import User from '../../model';
 
@@ -21,15 +20,19 @@ const client = new Client({
     partials: [Partials.Channel, Partials.Message],
 });
 
+interface CustomClient extends Client {
+    modals?: Collection<string, any>;
+}
+
 export const clientStart = async () => {
     try {
         await client.login(process.env.DISCORD_TOKEN);
+        (client as CustomClient).modals = new Collection();
     } catch (err) {
         const errMsg = getError(err);
         throw new Error(errMsg);
     }
 };
-discordModals(client);
 
 type cmdType = {
     name: string,
@@ -37,9 +40,9 @@ type cmdType = {
     options?: SlashCommandStringOption[],
 };
 
-const commandBuilder = (cmdArr: cmdType[]) => {    
+const commandBuilder = (cmdArr: cmdType[]) => { // builds commands
     const cmdArrBuilder: SlashCommandBuilder[] = [];
-    for (const command of cmdArr) {
+    for (const command of cmdArr) { // loops through commands
         const cmd = new SlashCommandBuilder()
             .setName(command.name)
             .setDescription(command.description);
@@ -86,17 +89,14 @@ type resType = {
     other?: boolean | undefined,
 }
 
-client.on('interactionCreate', async (interaction) => {
+
+client.on('interactionCreate', async (interaction) => { // slash command interaction
     try {
-        if (!interaction.isCommand() || !interaction.isModalSubmit) return;
-    
+        if (!interaction.isCommand() || !interaction.isModalSubmit ) return;
+
         const { commandName, options } = interaction;
-        
+
         switch (commandName) {
-            case 'ping':
-                await interaction.reply({embeds: [embedBuilder({title: 'Pong!', desc: 'Pong!'})]})
-            break;
-    
             case 'help':
                 await helpCmd(interaction);
             break;
@@ -110,12 +110,9 @@ client.on('interactionCreate', async (interaction) => {
             break;
     
             case 'check-follow':
-                const modal = await modalBuilder(interaction);
+                const modal = loginModalBuilder(interaction);
 
-                showModal(modal, {
-                    client: interaction.client,
-                    interaction: interaction,
-                })
+                await interaction.showModal(modal);
             break;
 
             case 'check-list':
@@ -131,57 +128,66 @@ client.on('interactionCreate', async (interaction) => {
     }
 });
 
-client.on('modalSubmit', async (i) => {
-    try {
-        const timeout = 210000; // 3.5 minutes 
 
-        let isTimedOut = true;
-    
-        const timeoutPromise = new Promise((resolve, reject) => { // timeout function, if discord command takes too long to complete, returns timeout err
-            setTimeout(() => {
-                if (isTimedOut) {
-                    reject(i.editReply({embeds: [embedBuilder({title: 'Error', desc: 'Request timed out. (you may have too many follows for the server to keep up)', err: true })]}));
-                }
-            }, timeout);
-        });
+client.on('interactionCreate', async (i) => { // modal submit interaction
+    if (i.type != InteractionType.ModalSubmit) return;
 
-        await Promise.race([ // runs both functions, if discord command takes too long to complete, returns timeout err
-            (async () => {
-                switch (i.customId) {
+    const timeout = 300000; // 5 minutes
+
+    let isTimedOut = true;
+
+    const timeoutPromise = new Promise((resolve, reject) => { // timeout function, if discord command takes too long to complete, returns timeout err
+        setTimeout(async () => {
+            if (isTimedOut) {
+                await i.editReply({embeds: [embedBuilder({title: 'Error', desc: 'Request timed out. (you may have too many follows for the server to keep up)', err: true })]});
+                reject(new Error('temp res'));
+            }
+        }, timeout);
+    });
+
+    await Promise.race([ // runs both functions, if discord command takes too long to complete, returns timeout err
+        (async () => {
+            try {
+       
+            switch (i.customId) {
                     case 'login':
-                        
-                        const username = i.components[0].components[0].value.trim();
-                        const password = i.components[1].components[0].value.trim();       
-                        
+                        const username = i.fields.getTextInputValue('username');
+                        const password = i.fields.getTextInputValue('password');
+
                         await i.reply({embeds: [embedBuilder({ title: 'Checking...', desc: `Checking follow list for ${username}. This may take a minute.` })]});
-                        
                         try {
                             const userData = new User({
                                 user_id: i.user.id, 
                             });
-    
+        
                             await userData.save();
+        
+                            await getFollowCmd(username, password, i);
+                            
                         } catch (err) {
+                            isTimedOut = false;
                             const errMsg = getError(err);
                             return i.editReply({embeds: [embedBuilder({ title: 'Error', desc: errMsg, err: true })]});
                         }
-
-                        await getFollowCmd(username, password, i);        
                         
                         await User.findOneAndDelete({ user_id: i.user.id });
 
-                        isTimedOut = false; // marks function as completed, doesnt run timeout func
+                        isTimedOut = false;
                     break;
-
                 }
-            })(),
-            timeoutPromise
-        ]);
 
-    } catch (err) {
+            } catch (err) {
+                isTimedOut = false;
+                const errMsg = getError(err);
+                throw new Error(errMsg);
+            }   
+        })(),
+        timeoutPromise
+    ]).catch((err) => {
         const errMsg = getError(err);
         throw new Error(errMsg);
-    }
+    });
 });
+
 
 export default client;
